[K/N] Fixed a corner case in casts optimization pass
The alogrithm handles differently two types of variables - T? and Boolean.
The first type is used in expressions like if (var == null) { .. } or if (var != null) { .. },
while the other is used in just if (variable) { .. } or if (!variable) { .. }. But in one case,
a variable might change its type - when a Boolean? variable gets aliased to a Boolean variable.
This wasn't expected and there was an assertion for such a case (a variable can't change its kind).
This commit fixes this by supporting this peculiar corner case.

 #KT-81257 Fixed