KT-81478 FileTreeWalkTest.withDirectoryFilter test
withDirectoryFilter filters out all directories
having a name that ends with "3".
Previously, root directory would always have a
".tmp" prefix, but after switching to createTempDirectory,
the directory name is a sequence of digits,
and the last one may be "3".

The fix changes a filtration condition in the test
to explicitly traverse a root directory.

^KT-81478 fixed