[K/N] Added a special "weak" edge into dependencies used for caches/linker
DependenciesTracker builds all sorts of dependencies for further usage (in codegen/linker/caches).
Incremental compilation also uses the result to find all the files for which caches are needed to be rebuilt.
There at least one case when this can be optimized: if a virtual function is called. If this function's declaring
interface/class has changed but with no change to this particular function signature (like, if a method has
been added), then this change on its own doesn't require to the dependencies' caches to be rebuilt. But the
actual cache is needed on the final stage - when invoking the linker.

To address this, a special new kind of edge is added to the dependencies graph - "weak" edge. Basically, it is
skipped during incremental compilation dependencies graph building but otherwise it works as usual edge.

 #KT-80929

(cherry picked from commit d19f7eea997cf9bc74e6b2019ecd8cdbf4fe6bd3)