fixup! [FIR] Stop reporting some `UPPER_BOUND_VIOLATED` from checkers
In this example, we first fix `V := SomeEnum` and then
`T := SomeInterface<SomeEnum>` as the only proper constraint.
Then, the constraint `T <: Enum<T>` becomes false as
`SomeInterface<SomeEnum> </: Enum<SomeInterface<SomeEnum>>`, but we
don't have any mechanism inside the constraints system for checking
that (and I fear we may not want to even, as there're probably some
scenarios that rely on the current behavior...)

Previously, the UBV checker would also check the fixation result as:
`SomeInterface<SomeEnum> <: Enum<SomeInterface<SomeEnum>> & SomeInterface<SomeEnum>`,
and during resolution we only add the constraints to the fresh variables
without any additional checks.

On the one hand, this could be a sign of a more generic issue with the
constraints system which we may want to handle separately, but on the
other we could argue that the logic that emits `WRONG_NUMBER_OF_TYPE_ARGUMENTS`
is there for exactly that reason...