[FIR] Redo less visible containing class checks
Basically, we drop the whole logic about containing classes and
instead check the type of the receiver expression.
This fixes some false positives when a public declaration declared in
a package private class is called through a public subtype (basically
only possible in Java, but nonetheless possible).

In case of a static call, the receiver expression is an
`FirResolvedQualifier` of the class through which the declaration is
called.
For Java declarations, it is again possible to call a static declaration
declared in a package private class through a public subclass.
This is considered legal by the checker because the bytecode will in
fact not contain any references to the package private super class, and
therefore it won't lead to access errors at runtime.

#KT-81262 Fixed