[kotlin] Do not consider the last indent of fenced KDoc code blocks as a part of these blocks
When there is a fenced code block (surrounded by ```), the fence has to be placed on a new line.
The spacing between the asterisk on this line and the closing fence was considered a part of this code block (`KDocTokens.CODE_BLOCK_TEXT`).
That's senseless, as no meaningful content can be placed there, otherwise, the code block is parsed incorrectly (in other editors / according to MD guides).
Now these spaces are viewed as `KDocTokens.TEXT`.

^KT-79783
^KT-75215