[K/N] Replace $ with \0 in InteropBridgesNameInventor to prevent clashes
InteropLowering emits name placeholders quoted with '$', e.g. `$name$`.
They are then parsed by InteropBridgesNameInventor.

The problem is that C code can legitimately contain '$' (see KT-81017
and KT-81538). So, when this happens, InteropBridgesNameInventor tries
to handle the following string as a name placeholder and fails.

This commit fixes that by using the null character instead of '$'.
The null character can't be used in valid C code, so no such clashes
should remain possible.

Using the null character is not ideal debug-wise, since it is not
displayed properly when printed.
But this choice seems like the only fully reliable one. After all, most
printable characters are allowed in C code at least in some places.

^KT-81017 Fixed
^KT-81538 Fixed