[K/N] Sanitize CCall.Direct symbol names in generated C stubs
To generate calls to C functions, the compiler generates and compiles
C stubs. In particular, when generating a call to a CCall.Direct
function, the compiler emits the specified symbol name to the generated
C stubs as a C string literal.

The problem is that it does that by simply adding quotes around the
symbol name and emitting it as is, without handling any special
characters that can't be used in string literals.
For example, `foo"bar` becomes `"foo"bar"`, which makes the generated
C code incorrect.

This commit fixes this by properly handling special characters.
Specifically, it leaves ASCII characters allowed in C string literals
by the C standard as is and replaces everything else with octal escape
sequences.

^KT-81548 Fixed