[FIR] Return null in `getSuperClassSymbolOrAny` in case of no runtime
There is no guarantee in Analysis API that a symbol for builtin
`kotlin.Any` will always be found. Hence, this assumption is removed
from the corresponding frontend utility function.
In case `kotlin.Any` is not encountered in some checker, it is best
not to report additional warnings given that the state is already
invalid.

^KT-78996 Fixed