[K/N] Worked around problem with deep nested loops
The algorithm tries to find a stable point during loops handling - this usually
takes several iterations (per loop) which leads to an exponential complexity -
(# of iteration)^(loops depth). Of course, such deep loops aren't common in practice,
but they might occur in the auto-generated code (this is the case for the library kotest).

This commit works around this by simply not handling functions which have deep nested loops.

 #KT-81617 Fixed