KLIB API: Introduce `Klib`, `KlibComponent` and  `KlibMetadataComponent`
The `Klib` interface represents an instance of Kotlin library (Klib)
that can be read by the compiler or any other tools.

It is supposed that in the future this interface will replace `KotlinLibrary`.

The `Klib` interface consists of multiple components, each responsible
for a certain aspect of the library. The first supported component
(and included in this commit) is `KlibMetadataComponent`, which provides
access to the metadata stored inside the library.

Other components (Ir, Ir of inlinable functions, Native-specific ones)
will follow in next commits.

^KT-81411