[K/N] Worked around a problem in CMP
When declaring a companion object of an ObjC class, it should be inherited from
the corresponding Meta class, but the compiler doesn't check this. But after
turning on safe casts (which check every type inconsistency in the IR and insert
a type check at runtime), it turned out, there are a lot of misuses of this. Leading
to runtime exceptions on previously working code.

This commit works around it by simply skipping the type checks for
companion objects of ObjC classes.

 #KT-81652 Fixed