FE: Avoid type variable fixation which leads to CST failure
As a fix for KT-74819, we've allowed semi-fixing of type variables
which only have other type variables as constraint (for input types)
that makes it possible to choose among OverloadResolutionByLambdaReturnType
annotated candidates.

But we can't find a resulting type for a variable which only has
lower constraints contained with other type variables, like for
Uv <: Tv, Fv <: Tv.
When we're trying to semi-fix Tv it leads to the computation of
a CST (common super type) of Uv, Fv, which is impossible and leads to
the failure (see the removed `.fir.fail`).

The idea is very straightforward: do not allow semi-fixing in such cases,
which might lead to the same OVERLOAD_RESOLUTION_AMBIGUITY as at KT-74819,
but only in very specific scenarios where it might be useful to ask
the user for specifying type arguments explicitly.

NB: It's not supposed to be a breaking change as affected
code would lead to the compiler internal error.

^KT-81254 Fixed