[Gradle] Use LazyResolvedConfigurationComponent in GMDT
Use LazyResolvedConfigurationComponent in platform-specific compilations
because GMDT doesn't need actual artifacts but rather
"component resolution result".

It also avoids problems when platform compilation can't get
resolved to certain artifact, i.e. artifact view with required artifactType
should be created. Android does that, they have "releaseCompileClasspath"
but when it hase dependency to another android project this configuration
can't be resolved directly. because dependency project would offer
a lot of secondary variants for different purposes, and Android consumer
aware of that. and use artifact view from that configuration
in all places when needed. KMP, unfortunately, is not aware of that
and thus fails with ambiguity when requesting artifacts.

But as previosly mentioned it is not needed for GMDT logic. So
replacing LazyResolvedConfigurationWithArtifacts with LazyResolvedConfigurationComponent
helps with beforementioned problem.

^KT-81060 Verification Pending