KLIB Metadata API: Drop `packageFragmentNames` parameter in `KlibMetadata*Factory*` classes
This parameter is not necessary to load outside `KlibMetadata*Factory*`
and then pass to one of the factories' methods. Because the list
of package fragment names can always be deduced inside the factory class.

This is a preparatory fix needed before landing the main changes
done in the scope of KT-81411.