KT-81693: introduce NativeKlibCompilationConfig interface
To be able to compile Native klibs without Kotlin/Native distribtution
we need to decouple the necessary parts from
the `kotlin-native/backend.native` module.

The crucial part of the compiler is `KonanConfig`: a huge wrapper around
CompilerConfiguration object and a couple of others. We likely want to
get rid of such thing completely. However, this would be a long and
complex thing to do.
Instead, we take a simpler approach here: decouple
the necessary parts of `KonanConfig` into `NativeKlibCompilationConfig`.

It is not the perfect abstraction, but it does this job. Later, after
complete move of klib compilation from the native backend, we will be
able to get rid of this interface completely.