KT-81693: Rename PhaseContext to NativeBackendPhaseContext
Naming things is hard. To decouple klib compilation from Native backend
we need to introduce a lighter version of PhaseContext which does not
depend on KonanConfig.
How should we call it? `LightPhaseContext`? Instead, this commit renames
`PhaseContext` to `NativeBackendPhaseContext` to highlight the purpose
of the existing interface, and allows us to use `PhaseContext` name for
a decoupled version.