KT-81703: Drop obsolete PSI to IR logic
Since 2.3.0 LV < 2.0 is unsupported. Which means we can safely remove
a huge pile of PsiToIr code. For a long time this was one of 2 hearts
of Kotlin/Native compiler alongside with IrToBitcode :)
Still, bits of the K1 frontend will be present in Kotlin/Native backend
for quite some time, as many parts (ObjCExport, C export, cinterop, etc)
still depend on descriptors.