FIR: fix secondary resolution of contracts with errors
The patched code in the `transformFunctionCallInternal` was intended
for the cases when error references are created with unresolved
arguments. In case described in the issue, the error reference to the
`contract` call already contains resolved (error) arguments, and
secondary resolution leads to the incorrect DFA processing.
The fix runs the resolutions only if arguments are not yet resolved.
#KT-80250 fixed

(cherry picked from commit 981855092ac4fb73eed609384379dff6419457de)