[K/N][build] Use -ffile-prefix-map in NativePlugin
To make clang output reproducible, all absolute paths must be
eliminated.

One solution is to pass every path as a relative path instead
of an absolute. But this is both insufficient and impossible.
Insufficient, because some information (e.g. debug information)
is still stored as absolute paths. Impossible, because some paths
are not stored relative to the project directory, but in some
global locations (which may also vary machine to machine).

Another solution, then, is the flag `-ffile-prefix-map=FROM=TO`,
which can be passed multiple times. Any path of the form
`<FROM>/some/path` will be present as `<TO>/some/path`
in the generated binary.

For `NativePlugin` it's now used for 3 substitutions:
- replace the root directory of the repo with `kotlin`
- replace the common root of all native dependencies with `NATIVE_DEPS`
- replace the jdk home with `JDK`, because many `NativePlugin` usages
  employ JNI, which uses headers from there.

^KT-81500