[K/N][build] Use -ffile-prefix-map in NativePlugin
To make clang output reproducible, all absolute paths must be
eliminated. `clang` is already called passing relative paths wherever
possible, but this is not enough.

A solution is the flag `-ffile-prefix-map=FROM=TO`, which can be passed
multiple times. Any path of the form `<FROM>/some/path` will be present
as `<TO>/some/path` in the generated binary.

For `NativePlugin` it's now used for 3 substitutions:
- replace the working directory with `.`
- replace the common root of all native dependencies with `NATIVE_DEPS`
- replace the jdk home with `JDK`, because many `NativePlugin` usages
  employ JNI, which uses headers from there.

^KT-81500