[Analysis API] Deprecate `KaFunctionType.arity`
This API had inconsistent behavior between K1/K2.
In K1, it represented the number from the ClassId
(e.g., `kotlin/Function3`), which represents all parameters
(including context parameters and receivers).
In K2, it represented just the number of value parameters.
Since both of these values are not that hard to compute manually,
it was decided to deprecate the API to avoid confusion.

^KT-80545 fixed