[Ir] Add `IrAnnotation`
Currently, annotations in IR are represented as `IrConstructorCall`s. This is too general representation, which should be replaced with specialized `IrAnnotation`. This commit is the first part of that change - it creates a new node (which, for temporary compatibility, is a subclass of `IrConstructorCall`), but does not yet introduce it into IR.

The next planned steps are:
1. Change the return type of `IrAnnotatedElement.annotations` from `List<IrConstructorCall>`` to `List<IrAnnotation>`
2. Detach `IrAnnotation` from `IrConstructorCall`.

^KT-74200