[FIR] Allow local objects to have type parameters again
This commit partially reverts changes made to fix KT-76806 which caused
KT-78985. Instead, it reimplements the fix on the checker side by not
assuming that the number of type arguments is equal to the number
of type parameters in the corresponding class symbol.

KT-78985 was never fully fixed because the analogous issue persisted
for inner objects (they also can have type parameters).
Some implementations might rely on the fact that local/inner classes
have all type parameters of containing callables/classes.

^KT-78985 Fixed