[LC] KTIJ-35990 Fix `psiType` to correctly map types like `SuspendFunction0` to jvm types
This in turn would allow `KotlinDeclarationInCompiledFileSearcher`
to properly find declarations with such suspend lambdas
as parameters. This would lead to decompiled light classes
for such declarations to have correct `kotlinOrigin`s, so that
they work correctly with the IDE's navigation and reference searches.

We cannot modify `JavaToKotlinClassMap` to do that for us,
documentation on it specifically states that it should not handle
numbered `SuspendFunction`s (see 1e41d144)

^KTIJ-35990 Fixed