fixup! [FIR] Stop creating synthetic function interfaces of arity >=255
Stop using `.can-freeze-ide` for the test: it only works with `PhasedJvmDiagnostic*`
runners and does so by accident, and other runners aren't even updated.
This change makes sure the LL runners and the one for the latestLV are
also generated. It's probably not worth adding any more complex support
for K1 shenanigans at this point, as we're going to get rid of it soon.