KT-81699: Move Native FIR to IR code to a new module
The new `:compiler:cli-native-klib` module contains parts of
`kotlin-native/backend.native/compiler/ir/backend.native` which are
relevant for compilation of sources into klib. This allows creating
native klibs without a native backend.

This refactoring is a bit clunky due to the strangeness of the
OutputFiles class. Still, this problem can be solved separately.

This commit completes the story of moving src -> klib logic from
the /kotlin-native directory. The new `:compiler:cli-native-klib` module
and `:compiler:ir.backend.native` should have enough "lego bricks"
for construction of a new CLI which can be done separately.

There are still problems with the current `NativeCompilerDriver`.
It should be moved to `:native:native-cli`. Such refactoring will make
 the Native backend a bit clearer. But this is out of the scope of KT-81699.