KT-81699: Move klib writer function to :native:fir-to-klib
This refactoring is a bit clunky due to the strangeness of the
OutputFiles class. Still, this problem can be solved separately.

This commit completes the story of moving src -> klib logic from
the /kotlin-native directory. The new `:native:fir-to-klib` module
and `:compiler:ir.backend.native` should have enough "lego bricks"
for construction of a new CLI which can be done separately.

There are still problems with the current `NativeCompilerDriver`.
It should be moved to `:native:native-cli` module which could potentially
depend on the future `native-klib-cli`. Such refactoring will make Native
backend a bit more clear. But this is out of the scope of KT-81699.