C-interop: Make the Klib writing endpoint independent of kotlinx-metadata-klib
The `createInteropLibrary()` function, which is the Klib writing
endpoint inside the C-interop tool, accepts `KlibModuleMetadata`
as the first parameter. `KlibModuleMetadata` is a part of
kotlinx-metadata-klib API.

With this commit, the endpoint starts using `SerializedMetadata`
instead of `KlibModuleMetadata`. Which makes the endpoint to
not depend on kotlinx-metadata-klib anymore. And makes it easier
to test.

^KT-81411