Reflection: fix equality for NothingKClass
`NothingKClass` delegates to `Void::class` to get all its behavior.
However, this delegation also unfortunately causes equals/hashCode to be
generated. Normally, custom equals/hashCode are not allowed in data
objects, but this is not checked in case implementations come from
interface delegation (KT-81751). The fix is to make `NothingKClass` not
`data`, and provide equals/hashCode explicitly, in which case interface
delegation will not generate overrides for them.

Also, implementation of `isNothingConstructor` in
`ReflectTypeSystemContext` was incorrect: it recognized `Void`'s type
classifier as a "nothing type", which led to it being a subtype of any
other type.

 #KT-81024 Fixed