[Test] Move the basic KlibLoader tests to the 'kotlin-klib-util' module
The source code of `KlibLoader` is located in the 'kotlin-klib-util'
module. However, the corresponding tests were located somewhere
in a compiler-related module.

Now, having the test fixtures feature, it's possible to move the base
`KlibLoader` test class to 'kotlin-klib-util'. And also, having
Klib mock DSL, it's possible to implement mock-based `KlibLoader` tests
directly in the same module.