[FIR] Deduplicate expect-actual pairs from binary dependencies during resolution
Actual declarations from binary dependencies don't preserve the isActual
flag.
Before this commit, this led to FirActualizingScope not filtering their
respective expects.
In lowPriorityActualFromDependeny.kt, we have an expect-actual pair
from binary dependencies with the actual one being annotated with
@LowPriorityInOverloadResolution.
This led to the actual declaration "losing" during resolution to a
candidate with a higher applicability, but the expect one being among
the "winners".
As a result, ConeEquivalentCallConflictResolver was not able to filter
out the expect declaration and OVERLOAD_RESOLUTION_AMBIGUITY was
reported.

To fix this, FirActualizingScope is adapted to match expect declarations
to non-expect ones using the same logic as
ConeEquivalentCallConflictResolver.

#KT-81592 Fixed