[Inline] Revert `c704685b396275a2c87bcd223093f8e865a52721`
Earlier we removed `allowExternalInlining` flag from the
`LoweringContext`. The idea behind this change was that we inline
any function if it is marked as inline and has a body. Everything else
(functions without bodies) is ignored.

In the mentioned commit we tried to fix an inlining problem by
reintroducing this flag. Probably the issue was located somewhere else.
After the revert, the issue is no longer reproducible, so the whole
commit can be removed.

^KT-80660 Fixed