[Gradle] Warn when configuration-on-demand is enabled for JS and Wasm
JS and Wasm do not support configuration on demand.
This can cause hard to diagnose errors from Gradle buildscripts that are not obviously caused by configuration-on-demand and JS/Wasm targets.

(Note that COD is still incubating, and is not fully supported by a significant proportion of Gradle plugins.)

Eventually we will have to update JS and Wasm so they are compatible with COD, but for now we can just log a warning to help prevent any confusion should unexpected errors occur.

^KT-79754