KLIB IO: Use the stable order of file tree traversal in zipDirAs()
The old implementation of zipDirAs() relied on the file system-dependent
file tree traversal order, which is not guaranteed to be stable.
In practice this means that the same uncompressed payload being
compressed in different environments may lead to different results,
if compared byte-to-byte.

The new implementation of zipDirAs() guarantees the same stable
order of traversal in different environments.

^KT-81709 Fixed