[ABI Validation] Renamed filtering blocks
After reviewing the existing filtering approach with exclude and include blocks, it was decided to keep it.

Although the logic of `include` filters is not obvious, this is a fairly common approach, both in Gradle and in other assembly systems and tools.
All other options are complicated and can confuse users and require them to explore a new filtering method.

In this commit, the filter block names were changed to `exclude` and `include` to match the names in Gradle and other tools.

#KT-80823 Fixed