[Inline] Add reproducer for KT-81903
In the given test we have two modules and enabled intra module inlining.
During compilation of the `main` module to klib, we are going to inline
function `foo` and copy without inlining function `ifTrue`
(from another module) together with its lambda. This lambda ends up
in a different file `InlineFunCallSite`
(originally from `InlineFunDeclaration`), and after the inlining
on the second stage we get incorrect file information for
this lambda because it was copied.

To get the correct stepping representation, remove the LANGUAGE
directive from the test (without removing the line).

#KT-81903