KLIB API: Split the hierarchy of Klib components
With this commit, we are introducing a new flavor of a Klib component:
`KlibMandatoryComponent`. It is a component that is always present
in the library.

`KlibOptionalComponent` stays in place. But now both mandatory and
optional components are inherited from the common ancestor:

                KlibComponent
                   /       \
KlibMandatoryComponent   KlibOptionalComponent

Also, the `getComponent()` accessors in `Klib` accept only
`KlibMandatoryComponent.Kind` or `KlibOptionalComponent.Kind`, but not
their common ancestor `KlibComponent.Kind` to avoid misuse:

interface Klib {
  fun getComponent(kind: KlibMandatoryComponent.Kind<KC>): KC
  fun getComponent(kind: KlibOptionalComponent.Kind<KC>): KC?
}

^KT-81411