[K/N] Switch custom cinterop to `-Xccall-mode both` by default
79bc1be introduces `@CCall.Direct` and the new flag to make cinterop
emit this annotation: `-Xccall-mode`.
It sets the flag to `indirect` by default (which disables emitting
`@CCall.Direct`), because other modes are not compatible with some
cinterop features: `-Xcompile-source` flag and `entryPoint` .def file
property.

Neither of the features seems to be widely used, so it is safe to change
the decision: flip the default mode to `both` and require users of those
incompatible features to switch it back explicitly.

With the `both` mode, cinterop klibs can be consumed in both modes --
through indirect and direct calls. But the compiler prefers indirect
calls by default, so this change doesn't affect the actual compilation
unless the compiler uses `-Xbinary=cCallMode=direct` or
`=directOrIndirect`. See 45cc5fe for details on this compiler flag.

In other words, switching cinterop `-Xccall-mode` to `both` by default
doesn't change the compilation by itself, but enables users consuming
such cinterop klibs with the new mode using the compiler flag. Once
more libraries in the ecosystem are compiled with `-Xccall-mode both`,
this will allow checking the applicability of the new mode at scale,
to discover and address problems that arise in practice.
This is the main goal of this change.

As a possible side effect, this could also allow the compiler reading
cinterop klibs produced by a newer Kotlin version (that uses a newer
LLVM version), because setting the compiler flag to `direct` would then
make it consume cinterop klibs without loading LLVM bitcode from them
(see 5706604).
But this is very tricky: the compiler flag is global (= affects all
cinterop klibs consumed by this compiler invocation), so if not
all dependencies are compatible with the new mode, it won't help.
And we also would need to do something with regular Kotlin klib ABI
compatibility, since cinterop klibs are only published along with
regular klibs. If the library author addresses that with the
`-language-version` flag, they could as well switch the cinterop klib to
the `direct` mode.

Another problem with the flag being global is that it also affects
platform libraries. And they are not compatible with the new mode yet.
For example, `errno` is not supported, but used by the core libraries.
So, this commit doesn't switch the mode for platform libraries (it
makes them use `-Xccall-mode indirect` explicitly), and doing that is
postponed (KT-82062).

^KT-81937 Fixed