[Gradle] replace legacy kotlin-test dependencies automatically to kotlin-test
For historical reasons we had in the past:
* kotlin-test-common
* kotlin-test-annotations-common

that had to be used in common source sets to unlock syntax highlighting
and metadata compilations.
This is no longer the case for a long time. And in fact it can cause
problems when metadata compilation will be enabled for kotlin-test.

Because for this dependency structure:

```
commonTest -> kotlin-test-common
jvmTest -> kotlin-test-junit
```

Kotlin metadata compiler ignores kotlin-test-common, and don't
granular metadat dependency transformation will not be able to infer
that kotlin-test is present, as it works correctly only when
dependency cooridnates is consistent across all targets i.e.
kotlin-test should be present in resolution resoluts for
metadata & platform compilations.

To solve that, these changes introduces the automatic replacement
i.e. kotlin-test will be present when `kotlin-test-common` or
`kotlin-test-annotation-common` is added on any source set.

Later, it would be nice to nudge users to replace them explicitly.

Also added more kotlin-test to verify various cases.

^KT-81849 Verification Pending