Reflection: make CapturedKType a subclass of AbstractKType
Otherwise there's no way to represent a flexible captured type
(`FlexibleKType` needs `AbstractKType` lower/upper bounds), which is
possible, for example, for types from Java hierarchies.

Most properties available in `AbstractKType` do not make sense for
captured types and/or are only used in `toString` or other functions
where it's impossible to observe a `CapturedKType` instance. The only
exceptions are `makeNullableAsSpecified` (which was previously supported
indirectly in `ReflectTypeSystemContext`), and
`isDefinitelyNotNullType`/`makeDefinitelyNotNullAsSpecified` (but
captured types are not DNNs because their classifier is not a type
parameter).

 #KT-81967 Fixed