[K/JS] Don't use defaultType inside TestGenerator for promiseSymbol
In some cases (which I haven't found to be honest),
the Promise symbol is unbound inside the TestGenerator lowering.
I've tried multiple setups but haven't reproduced the behavior
filed in ^KT-79514.

To unblock users, I've dropped usage of the `defaultType`,
in the tests lowering, however, it's definitely not the root cause.


Merge-request: KT-MR-24206
Merged-by: Artem Kobzar <Artem.Kobzar@jetbrains.com>