KLIB API: Make component loading to be lazy
With this commit, Klib components are no more created and registered
during the creation of a `Klib` instance. Now they are created
at the call site where and when the component is accessed for
the first site. And each created component is 'cached' in a Klib
instance for further reuse.

Example:
  val klib: Klib = ...
  // ^^^ Klib is created here with the empty set of components
  ...
  val metadata = klib.metadata
  // ^^^ The first access to `KlibMetadataComponent` leads to
  // creating and registering this component in `klib`

This also brings a flexibility w.r.t. custom components: A custom
component located somewhere in a user's module can be used
as transparently as the default components.

Example:
  // component located in user's module A:
  interface MyComponent : KlibComponent {
    companion object Kind : KlibComponent.Kind<...>
  }

  // component usage in user's module B that depends on module A:
  val klib: Klib = ...
  // ^^^ Here there is no `MyComponent` yet
  ...
  val my = klib.getComponent(MyComponent.Kind)
  // ^^^ And here it is created and cached in the `klib` instance

^KT-81411