[K/N] Support generating `@CGlobalAccess` in cinterop
See KT-79742 for the context.

79bc1be introduced a new flag `-Xccall-mode` to cinterop which made it
generate `@CCall.Direct` annotations instead of/in addition to `@CCall`,
instructing the compiler how to call those C functions directly, without
using the generated wrappers.

This commit does similar thing for C globals, using the same flag as
the mode selector:
* `-Xccall-mode indirect` disables generating the new annotations
* `-Xccall-mode both` adds `@CGlobalAccess(name)` to property accessors
* `-Xccall-mode direct` adds `@CGlobalAccess(name)` and removes `@CCall`

In addition, certain C globals are represented in Kotlin in such a way
that Kotlin property value is not a value of the global, but a pointer
to the global. This is true for C array- and struct-type globals.
To make this distinction known to the compiler, cinterop marks accessors
(= getters) of such globals with another new annotation,
`@CGlobalAccess.Pointer`.

^KT-79752