[BTA] Deprecate JvmSnapshotBasedIncrementalCompilationOptions
Also deprecate the constructor for
JvmSnapshotBasedIncrementalCompilationConfiguration as we will require
instantiation through a factory method, like all other types in BTA, to
help us evolve the internal implementation.

We don't really need two objects with nesting to hold the
data, as JvmSnapshotBasedIncrementalCompilationConfiguration itself is
enough.

This is part of preparations for making operations immutable.

^KT-81790