FE/fixation: rewrite `VariableFixationFinder.getReadiness()`
This is the initial attempt to rewrite `getReadiness()` from a single
enum into a vector with lexicographical comparison.

The benefit of this approach is that now the different aspects we
consider when we compare type variables have the option to remain
independent from one another - and it is actually the desired state of
thigs for us as it "preserves the details". A variable with a proper
non-ILT equality constraint will always remain more prioritized to a
variable without one but with all the same other qualities regardless
what those qualities are. Previously, concluding that a variable has
some readiness other than `WITH_COMPLEX_DEPENDENCY` would mean that this
nuance is completely lost.

The new approach allows both replicating the old one precisely and
keeping the qualities independent, so this commit aims to make a
"half-step" in the desired direction while mainaining the original
code's flow by using quite complex conditions and smelly `if` checks
(like the one with `HAS_NO_DEPENDENCIES_TO_OTHER_VARIABLES`, which
disables the subsequent component altogether - this is what our previous
`when` branches used to do, basically, if we view them through the lens
of the new approach).

Note that the subsequent commits adjust the code further to achieve a
simpler final version.

^KT-77939