[FIR] Continue resolution after inaccessible receiver in secondary constructor
Inside secondary constructor parameters and delegated constructor calls,
we had a resolution behavior that was inconsistent with primary
constructors.
We could resolve to members inside the class but would always report
INSTANCE_ACCESS_BEFORE_SUPER_CALL instead of continuing the resolution
to possibly applicable candidates.

In KT-15152, we introduced InaccessibleReceiverKind.ClassHeader which
does not produce applicable candidates, therefore not forcing the
tower resolution to stop when encountering such a candidate.

When ImprovedResolutionInSecondaryConstructors is enabled, we stop
adding an InaccessibleImplicitReceiverValue with kind
SecondaryConstructor to the tower context in secondary constructor
parameters and delegated constructor calls.
Secondary constructors use the same tower context as the class
header, so we still report INSTANCE_ACCESS_BEFORE_SUPER_CALL if we don't
find any applicable candidates.

#KT-77275 Fixed
#KT-77276 Fixed