[K/N] Add a pointer to `KotlinObjCClassInfo` into `KotlinObjCClassData`
Kotlin classes can extend Objective-C classes. Such classes are called
"Kotlin Objective-C classes".

`KotlinObjCClassInfo` is data generated by the compiler for such a
class, used to create a peer Objective-C class at runtime.

Such created Objective-C classes have a pointer to `KotlinObjCClassData`
-- data statically allocated by the compiler but filled at runtime and
used for implementation of certain methods for such Objective-C classes.

KT-79741 aims to provide an alternative, "direct" implementation for
type checks against Objective-C protocols in the compiler.
As explained in the next commits, to achieve that, the runtime needs to
find `KotlinObjCClassInfo` from a created Objective-C class.
Since such a class already has a pointer to `KotlinObjCClassData`, it is
enough to make `KotlinObjCClassData` have a pointer to
`KotlinObjCClassInfo`.

This commit does that: it adds the new field and makes the compiler
fill it.

^KT-79741