[K/N] Mark runtime type info for Kotlin Objective-C classes with a flag
Kotlin classes can extend Objective-C classes. Such classes are called
"Kotlin Objective-C classes".

KT-79741 aims to provide an alternative, "direct" implementation for
type checks against Objective-C protocols in the compiler.
As explained in the next commits, to achieve that, the runtime needs an
efficient way of distinguishing Kotlin objects that are instances of
Kotlin Objective-C classes.

Kotlin objects have a pointer to `TypeInfo`, and `TypeInfo` has flags.
So this commit makes the runtime mark the corresponding `TypeInfo`s with
a new flag, `TF_KOTLIN_OBJC_CLASS`.

The mind-blowing part is that Kotlin Objective-C classes actually have
two `TypeInfo`s:
* One is created by the compiler, but it is "fake".
* Another is created at runtime and is actually used in Kotlin objects.
  The runtime uses the created Objective-C class and the first Kotlin
  `TypeInfo` as input to create this second `TypeInfo`.

Naturally, only the latter needs to be marked with the flag. Moreover,
both `TypeInfo` for the created Objective-C class and meta-class need to
be marked, since both require type-checking working.

Luckily, there is already (an inefficient) way of distinguishing such
Objective-C classes, and this commit makes the runtime use it to set the
flag when it creates `TypeInfo`s.

^KT-79741